增加查询接口

+++ data-jpa/src/main/kotlin/top/bettercode/summer/data/jpa/web/MybatisPagedResources.kt
@@ -10,18 +10,18 @@ class MybatisPagedResources<T>(
         /**
          * 当前页
          */
-        @field:JsonView(Any::class) var page: Long,
+        @field:JsonView(Any::class) var page: Int,
         /**
          * 每页显示条数，默认 10
          */
-        @field:JsonView(Any::class) var size: Long,
-        @field:JsonView(Any::class) var pages: Long,
+        @field:JsonView(Any::class) var size: Int,
+        @field:JsonView(Any::class) var pages: Int,
         /**
          * 总数
          */
         @field:JsonView(Any::class) var total: Long,
         @field:JsonView(Any::class) var list: Collection<T>) {
-    constructor(page: Page<T>) : this(page.number.toLong(), page.size.toLong(), page.totalPages.toLong(), page.totalElements,
+    constructor(page: Page<T>) : this(page.number, page.size, page.totalPages, page.totalElements,
             page.content)
 
 }
