+++ web/src/main/kotlin/top/bettercode/summer/web/PagedResources.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonView
 import org.springframework.util.Assert
 import java.util.*
 
-class PagedResources<T>(number: Long, size: Long, totalPages: Long, totalElements: Long, content: T) {
+class PagedResources<T>(number: Int, size: Int, totalPages: Int, totalElements: Long, content: T) {
     @get:JsonView(Any::class)
     var content: T? = content
 
@@ -16,11 +16,13 @@ class PagedResources<T>(number: Long, size: Long, totalPages: Long, totalElement
         this.page = PageMetadata(number, size, totalPages, totalElements)
     }
 
-    class PageMetadata(@JsonView(Any::class)
-                       @JsonProperty var number: Long, @JsonView(Any::class)
-                       @JsonProperty var size: Long, @JsonView(Any::class)
-                       @JsonProperty var totalPages: Long, @JsonView(Any::class)
-                       @JsonProperty var totalElements: Long) {
+    class PageMetadata(
+        @JsonView(Any::class)
+        @JsonProperty var number: Int, @JsonView(Any::class)
+        @JsonProperty var size: Int, @JsonView(Any::class)
+        @JsonProperty var totalPages: Int, @JsonView(Any::class)
+        @JsonProperty var totalElements: Long
+    ) {
 
         init {
             Assert.isTrue(number > -1, "Number must not be negative!")
@@ -30,8 +32,10 @@ class PagedResources<T>(number: Long, size: Long, totalPages: Long, totalElement
         }
 
         override fun toString(): String {
-            return String.format("Metadata { number: %d, total pages: %d, total elements: %d, size: %d }", number,
-                    totalPages, totalElements, size)
+            return String.format(
+                "Metadata { number: %d, total pages: %d, total elements: %d, size: %d }", number,
+                totalPages, totalElements, size
+            )
         }
 
         override fun equals(other: Any?): Boolean {
